Safari-611.1.10/LayoutTests/http/tests/xmlhttprequest/post-blob-content-type-async-expected.txt

CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-allow-lists.php due to access control checks.
Test verifies that content MIME type is set correctly when Blob is sent using XMLHttpRequest asynchronously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS expectedMimeType is "text/plain;charset=utf-8"
PASS expectedMimeType is ""
PASS expectedMimeType is ""
PASS expectedMimeType is ""
PASS expectedMimeType is ""
PASS expectedMimeType is ""
PASS expectedMimeType is ""
PASS expectedMimeType is "multipart/mixed;boundary=\"--blob-boundary"
PASS expectedMimeType is "multipart/mixed;boundary=\"--blob-boundary\""
PASS expectedMimeType is ""
PASS expectedMimeType is ""
PASS Cross-origin request without CORS headers should fail.
PASS expectedMimeType is "text/plain;charset=utf-8"
PASS successfullyParsed is true

TEST COMPLETE
